breaking out the examples into multiple files
